Master scheduler is no more; logic moved to kernel
